pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Joshua Szepietowski", email = "joshSzep@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.6",
    "gunicorn>=23.0.0",
    "uvicorn>=0.34.0",
    "channels>=4.2.0",
    "websockets>=14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "pytest-asyncio>=0.25.3",
    "pyright>=1.1.394",
    "pre-commit>=4.1.0",
    "pytest-django>=4.10.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
force-single-line = true
combine-as-imports = false
split-on-trailing-comma = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.pyright]
pythonVersion = "3.13"
reportMissingImports = "error"
reportUnusedImport = "error"
reportImportCycles = "error"

[[tool.pyright.executionEnvironments]]
root = "."
extraPaths = [
    "./src",
    "./test",
    "./",
]
